jangbwaguni/requirements.txt
asgiref==3.4.1
Django==3.2.9
django-multiselectfield==0.1.12
django-phonenumber-field==6.0.0
djangorestframework==3.12.4
phonenumberslite==8.12.37
Pillow==8.4.0
psycopg2==2.9.2
pytz==2021.3
sqlparse==0.4.2
